Move preprocess_f0 func to preprocess.py and melodyU encoder consist of fft encoder. I Implemented fft.py at pr #24.
